# [Refactoring] Improve spline/bezier.