--- a/GPU/Common/DrawEngineCommon.h
+++ b/GPU/Common/DrawEngineCommon.h
@@ -178,6 +178,5 @@ protected:
 	bool fboTexBound_ = false;
 
 	// Hardware tessellation
-	int numPatches;
 	TessellationDataTransfer *tessDataTransfer;
 };