… GPU/Common/SplineCommon.h
… GPU/Common/VertexDecoderCommon.h
… GPU/D3D11/DrawEngineD3D11.cpp
… GPU/D3D11/DrawEngineD3D11.h
… GPU/Directx9/DrawEngineDX9.cpp
… GPU/Directx9/DrawEngineDX9.h
… GPU/Directx9/VertexShaderGeneratorDX9.cpp
… GPU/GLES/DrawEngineGLES.cpp
… GPU/GLES/DrawEngineGLES.h
… GPU/GLES/GPU_GLES.cpp
… GPU/GLES/ShaderManagerGLES.cpp
… GPU/GLES/ShaderManagerGLES.h
… GPU/GLES/VertexShaderGeneratorGLES.cpp
… GPU/GPUCommon.cpp
… GPU/GPUState.h
… GPU/Math3D.cpp
… GPU/Math3D.h
… GPU/Software/TransformUnit.cpp
… GPU/Vulkan/DrawEngineVulkan.cpp
… GPU/Vulkan/DrawEngineVulkan.h
… GPU/Vulkan/VertexShaderGeneratorVulkan.cpp
… UI/GameSettingsScreen.cpp
… Windows/GEDebugger/VertexPreview.cpp
… ext/native/thin3d/GLQueueRunner.cpp
… ext/native/thin3d/GLQueueRunner.h
… ext/native/thin3d/GLRenderManager.h