This is a little big rewrite of spline/bezier include many improvements, bug fixes and refactoring.

#### Performance improvements

- Improve the speed of software tessellation approximately 15-25 fps on PC/Android. (For example, FF4 thundaga on Android: 5 fps to 20 fps)
- Hardware tessellation worked on my "weak" GPU(adreno 305), also the speed improved approximately 40 fps(20 fps to 60 fps) with B-spline demo(bezier in pspgl).
- [Edit] Unfortunately HW tessellation is not so fast on weak GPU like old mobile phone. It seems just get almost same speed as SW tessellation. Anyway SW tessellation seems faster enough for now.
- [Edit] Pursuit Force is a bit slower when using HW tessellation on GLES.

#### Feature improvements

- Get rid of divisions in spline weights calculation.
- Use derivatives for B-spline normals generation.
- Use same weights if u == v to avoid duplicated calculations.
- Use cached weights.
- Pre-convert the control points to avoid conversion in hot loops.
- Use real tessellation for low-quality.
Before(spline/bezier is same geometory)
![bezi00678_00000](https://user-images.githubusercontent.com/1607422/46157873-a520e400-c2b7-11e8-98d1-b541e518605e.jpg)
![bezi00678_00001](https://user-images.githubusercontent.com/1607422/46157874-a520e400-c2b7-11e8-82a0-7261b2893e98.jpg)
After(Bezier)
![bezi00678_00000](https://user-images.githubusercontent.com/1607422/46158247-67708b00-c2b8-11e8-9ce2-c11374eb1748.jpg)
![bezi00678_00001](https://user-images.githubusercontent.com/1607422/46158248-68092180-c2b8-11e8-9cc3-5f03076276fc.jpg)
High quality(bezier)
![bezi00678_00002](https://user-images.githubusercontent.com/1607422/46158039-ffba4000-c2b7-11e8-9ea3-109d86315bd7.jpg)
After(Spline)
![bezi00678_00002](https://user-images.githubusercontent.com/1607422/46158263-75bea700-c2b8-11e8-8145-a8a623bd0f03.jpg)
![bezi00678_00003](https://user-images.githubusercontent.com/1607422/46158264-75bea700-c2b8-11e8-8b5e-902a3b98e5fd.jpg)
High quality(spline)
![bezi00678_00003](https://user-images.githubusercontent.com/1607422/46158037-ff21a980-c2b7-11e8-9ddb-2ed8b2c25ce3.jpg)
- Some SIMD optimizations(correctly optimized?).
- Managing a raw memory buffer.
- Add support for changing the quality of hardware tessellation.
- Disable instanced tessellation.

#### Bug fixes

- Fix texture coordinates of low-quality.
Before
![ulus10560_00000](https://user-images.githubusercontent.com/1607422/46159201-a9023580-c2ba-11e8-989d-b85694ed6db5.jpg)
After
![ulus10560_00008](https://user-images.githubusercontent.com/1607422/46159461-55dcb280-c2bb-11e8-80d5-29f0c11c4f16.jpg)
#### Refactoring

- Unify spline/bezier tessellation logic using template.

#### Known issue [Edit]

- SplineSurface demo(from jpcsp) is something wrong with SW tessellation. However, HW tessellation seems correct. This issue is known since before these commits.
HW tess
![spli01764_00001](https://user-images.githubusercontent.com/1607422/46208586-2c2c9580-c366-11e8-9cd6-0f0adc1282f1.jpg)
SW tess
![spli01764_00004](https://user-images.githubusercontent.com/1607422/46208587-2cc52c00-c366-11e8-839b-c2c030b30342.jpg)

#### Future plan(TODO) [Edit]

- Implement HW tessellation for DX9.
- Hand optimization with ARM NEON if I can.

I just wrote something that is remembered for now, so I may add the description later.
Sorry for the huge pull request.