@hrydgard 2018-09-27
Whoa. Won't have time to look at this until the weekend but sounds promising :)