@marosis 2018-09-27
Owhdusbduxbfzykqhxkabw WHAAAAAAAT? You have JUST REALLY said FPS improve over 15-25???????? I don't have words