@LunaMoo 2018-09-28
Nice overall, but hardware tesselation seems to still be buggy in "Juiced 2: Hot Import Nights" so this also breaks it when spline/bezier curves are set to low now. The game uses curves to draw road and since it's a racing game people with low end hardware might abuse setting the curves to low here. Anyway the glitch of hardware should be easily noticeable in this dump:

[ULES00928_0001.zip](https://github.com/hrydgard/ppsspp/files/2426551/ULES00928_0001.zip)

And probably shouldn't be a blocking case for this PR since hardware tesselation is still buggy same way on master.