@unknownbrackets 2018-09-28
Just wanted to say these performance improvements look very cool.

But @marosis note that it only affects certain games / certain areas of games.  Most of the FPS notes are using tests that render only spline/bezier.  Don't expect it to make every game way faster - it's a very cool update that will make some games way faster, though, so it's still something to be excited about.

-[Unknown]