@Emulatorer 2018-09-28
Maybe most of the improvements are only working in a android phone that supports HW Tesselation (Adreno GPU)  I think :/