@xebra 2018-09-28
Thanks all.
@hrydgard @unknownbrackets Sorry for the huge commits in one time. I kept in mind to simple, clean, clear, easy to read, so I hope you can read easily whole code.

@LunaMoo I'm not sure what is the problem about hw tess. Also I don't know how can I get any information from the dump file. Thanks anyway to some tests.

@marosis I'm sure you can get any effects when you play games in very heavy spline/bezier situation.