@unknownbrackets 2018-09-28
To use a GE dump - just open it like a game.  I usually drag them into the PPSSPP window.

It will run all the same GE commands that the game sent for that frame, just over and over again each frame.  Note that anything that was render-to-texture or any block transfers will now come from RAM, so it's not perfect, but it does get you most of the frame behavior.

-[Unknown]