@zminhquanz 2018-09-29
@xebra Hello , long time no see , i'll expect about this . How about HW Tesellation on GLES 2 and DX9 ?