@weihuoya 2018-09-29
oh, Pursuit Force use a lot of bezier, Metal Gear Acid 2 use a lot of spline, and will get performance improvement.