@xebra 2018-09-29
@zminhquanz Hi! I will try to implement for DX9 after this. Also, GLES2 is... I need more research on it.