@xebra 2018-09-29
BTW, I have a question.
We call Flush() twice soon in spline/bezier.
This seems wasted. So, can we get rid of one of these?

```
void GPUCommon::Execute_Bezier(u32 op, u32 diff) {
	drawEngineCommon_->DispatchFlush();
	...
	drawEngineCommon_->SubmitBezier(...);
}
...
void DrawEngineCommon::SubmitBezier(...) {
	PROFILE_THIS_SCOPE("bezier");
	DispatchFlush();
	...
}
```