@unknownbrackets 2018-09-29
Well, the first flush is in case we have other prims/triangles/etc., which can't be combined with the bezier.  The second one is to draw the actual bezier because it can't be combined with later drawing.

-[Unknown]