@xebra 2018-09-29
Really?
Actually, Flush() is called three times in this case because it's called twice in SubmitBeizer().

void DrawEngineCommon::SubmitBezier(...) {
	PROFILE_THIS_SCOPE("bezier");
	DispatchFlush();
	...
	// Tessellation...
	...
	DispatchSubmitPrim(...);

	DispatchFlush();

	if (origVertType & GE_VTYPE_TC_MASK) {
		gstate_c.uv = prevUVScale;
	}
}

I think you said  "The second one is to draw the actual bezier because it can't be combined with later drawing." is meaning last one.