@xebra 2018-09-29
Pursuit Force on Android seems almost don't gain performance improvements.
Probably, it's not bezier's issue, it's because of heavy(approx. 1000) draw calls(4x4 bezier patch) in one frame, I think.