@hrydgard 2018-09-29
For Pursuit Force we might then want to try to combine draw calls even with beziers.. Should be possible, just need a little extra state tracking I guess..