@weihuoya 2018-09-29
I had try to combine beziers and splines drawcalls, it do not have any improvement.