@xebra 2018-09-30
I guess GLES2 devices have old and weak GPU, so even if we can support GLES2 for HW tess, maybe it's slower than SW tess(CPU).
Also If we decide to support GLES3 or higher, maybe we can use more strong GL features.
How about it?