@unknownbrackets 2018-09-30
I personally think these improvements are great, and GLES2 (using software) will already benefit a lot from them.  I also think GLES2 might perform worse with hardware (even my old NVIDIA desktop GPU did, iirc.)

I tested this with a couple games, but starting to need to keep a database of GE dumps with tags for things they do... I can't remember which games that I have use bezier/spline anymore.

Anyway, seems great so far from what I've tried.

-[Unknown]