@xebra 2018-09-30
@unknownbrackets Thank you for testing these. Your opinion is helpful to me.

Current HW tess on OpenGL is a little slower than Vulkan/DX11. It's probably because of using texture as tessellation data transfer. So, if we can use any generic buffer like TBO/UBO instead texture, maybe we can boost speed more.