@zminhquanz 2018-09-30
@xebra if HW Tessellation on GLES 2 are too slow , you can make a check to enable SW Tessellation on these