@LunaMoo 2018-10-01
I think the option of hardware tesselation should be renamed to something like  "hardware spline/bezier curves" or "generate spline/bezier curves on gpu", as seeing the above comment and similar confusion in the past on the forums, people still think this is some kind of graphic enhancement or a speedhack that they absolutely have to use. While actually it's not even as compatible as software curves yet and might reduce quality in affected games, performance might also be pretty terrible for example on amd gpu's or older gpu's in general compared to software curves:].