@hrydgard 2018-10-01
Agree that renaming it makes sense. Or removing it and autodetect whether to use it somehow... although that might be a bit unrealistic still.