@Emulatorer 2018-10-01
@xebra this could help #10842 or not?