@unknownbrackets 2018-10-02
I don't think that game uses splines/beziers, but if it does it could help.

-[Unknown]