@zminhquanz 2018-10-02
@xebra I have a one thing . If DX9 and GLES 2 isn't use HW tess , using SW Tess is ok because HW tess isn't improve performance on these API . If you want to use HW tess , using GLES 3 and modern API