@unknownbrackets 2018-10-06
Just a note: with #11447 merged, we'll want to delete texture subimage data if necessary here:
https://github.com/hrydgard/ppsspp/pull/11447#discussion_r223186670

-[Unknown]