@xebra 2018-10-07
I tried fixing it. Thanks to the note @unknownbrackets .