@hrydgard 2018-11-04
I'm happy with it too - let's go for it.