@Emulatorer 2018-11-05
@xebra what are those games that really benefit this Refactoring and even my doesn't support HW Tesselation I can feel a little bit improvement in the performance? 

Just Asking :)