@xebra 2018-11-05
Thank you for approval and merging. I'm happy that these were merged.
And, now I'm trying to fix the low-quality issue but it needs more time.

@Emulatorer Only on games using Spline/Bezier.
I don't know what game is using these but I just testing with FF4, Loco Roco, Pursuit Force, and some demos.
I think you should search in this issue/PR lists, and then you can get some information about those.
Please don't get your hopes up about HW Tessellation because this is still a highly experimental feature.